+++ lib/hexapdf/configuration.rb
@@ -624,6 +624,7 @@ module HexaPDF
                         dereference: 'HexaPDF::Task::Dereference',
                         pdfa: 'HexaPDF::Task::PDFA',
                         merge_acro_form: 'HexaPDF::Task::MergeAcroForm',
+                        import_pages: 'HexaPDF::Task::ImportPages',
                       })
 
   # The global configuration object, providing the following options: